Abundance density plot

Density abundance profile of single features in a TreeSummarizedExperiment. The panel implements plotAbundanceDensity to generate the plot.

Value

The AbundanceDensityPlot(...) constructor creates an instance of an AbundanceDensityPlot class, where any slot and its value can be passed to ... as a named argument.

Slot overview

The following slots control the thresholds used in the visualization:

layout, a string specifying abundance layout (jitter, density or points).

assay.type, a string specifying the assay to visualize.

n, a number indicating the number of top taxa to visualize.

flipped, a logical specifying if the axis should be switched.

order_descending, a string specifying the descending order.

In addition, this class inherits all slots from its parent Panel class.

Author

Giulio Benedetti

Examples

# Import TreeSE
library(mia)
#> Loading required package: TreeSummarizedExperiment
#> Loading required package: Biostrings
#> Loading required package: XVector
#> 
#> Attaching package: ‘Biostrings’
#> The following object is masked from ‘package:base’:
#> 
#>     strsplit
#> Loading required package: MultiAssayExperiment
data("Tengeler2020", package = "mia")
tse <- Tengeler2020

# Add relabundance assay
tse <- transformAssay(tse, method = "relabundance")

# Launch iSEE
if (interactive()) {
  iSEE(tse)
}