Graph plot

The Graph plot is a virtual class that showcases the network organisation of either the features or samples of a SummarizedExperiment object. The RowGraphPlot and ColumnGraphPlot classes belong to this family and are specialised to visualise the feature or sample igraphs stored in metadata, respectively.

Slot overview

The following slots control the thresholds used in the visualisation:

name: Character scalar. Metadata name containing graph. (Default: "graph")

assay.type: Character scalar. Assay type to use. (Default: "counts")

layout: Character scalar. Graph layout. (Default: "kk")

edge.type: Character scalar. Edge type. (Default: "fan")

show.label: Logical scalar. Should node labels be shown. (Default: FALSE)

add.legend: Logical scalar. Should legend be shown. (Default: TRUE)

edge.colour.by: Character scalar. Parameter to colour lines by when colour_parameters = "Edge". (Default: NULL)

edge.size.by: Character scalar. Parameter to size lines by when size_parameters = "Edge". (Default: NULL)

node.colour.by: Character scalar. Parameter to colour nodes by when colour_parameters = "Node". (Default: NULL)

node.size.by: Character scalar. Parameter to size nodes by when size_parameters = "Node". (Default: NULL)

node.shape.by: Character scalar. Parameter to shape nodes by when shape_parameters = "Node". (Default: NULL)

In addition, this class inherits all slots from its parent class Panel.

See also

Author

Giulio Benedetti