RDA plot

CCA/RDA plot for the rows of a TreeSummarizedExperiment object. The reduced dimension can be produced with runRDA and gets stored in the reducedDim slot of the experiment object. The panel implements plotRDA to generate the plot.

Value

The RDAPlot(...) contructor creates an instance of a RDAPlot class, where any slot and its value can be passed to ... as a named argument.

Slot overview

The following slots control the thresholds used in the visualization:

add.ellipse, a string specifying ellipse layout (filled, coloured or absent).

colour_by, a string specifying the parameter to color by.

add.vectors, a logical indicating if vectors should appear in the plot.

vec.text, a logical indicating if text should be encased in a box.

confidence.level, a numeric between 0 and 1 to adjust confidence level.

ellipse.alpha, a numeric between 0 and 1 t o adjust ellipse opacity.

ellipse.linewidth, a numeric specifying the size of ellipses.

ellipse.linetype, a numeric specifying the style of ellipses.

vec.size, a numeric specifying the size of vectors.

vec.colour, a string specifying the colour of vectors.

vec.linetype, a numeric specifying the style of vector lines.

arrow.size, a numeric specifying the size of arrows.

label.colour, a string specifying the colour of text and labels.

label.size, a numeric specifying the size of text and labels.

add.significance, a logical indicating if variance and p-value should appear in the labels.

add.expl.var, a logical indicating if variance shoud appear on the coordinate axes.

In addition, this class inherits all slots from its parent Panel class.

Author

Giulio Benedetti

Examples

# Import TreeSE
library(mia)
data("enterotype", package = "mia")
tse <- enterotype

# Run RDA and store results into TreeSE
tse <- runRDA(tse,
              formula = assay ~ ClinicalStatus + Gender + Age,
              FUN = vegan::vegdist,
              distance = "bray",
              na.action = na.exclude)

# Launch iSEE with custom initial panels
if (interactive()) {
  iSEE(tse, initial = c(RDAPlot()))
}

#' @docType methods