Row tree plot

Hierarchical tree for the rows of a TreeSummarizedExperiment object. The tree can be produced with addTaxonomyTree and gets stored in the rowTree slot of the experiment object. The panel implements plotRowTree to generate the plot.

Value

The RowTreePlot(...) constructor creates an instance of a RowTreePlot class, where any slot and its value can be passed to ... as a named argument.

Slot overview

The following slots control the thresholds used in the visualization:

layout, a string specifying tree layout

add_legend, a logical indicating if color legend should appear.

edge_colour_by, a string specifying parameter to color lines by when colour_parameters = "Edge".

edge_size_by, a string specifying parameter to size lines by when size_parameters = "Edge".

tip_colour_by, a string specifying parameter to color tips by when colour_parameters = "Tip".

tip_size_by, a string specifying parameter to size tips by when size_parameters = "Tip".

tip_shape_by, a string specifying parameter to shape tips by when shape_parameters = "Tip".

node_colour_by, a string specifying parameter to color nodes by when colour_parameters = "Node".

node_size_by, a string specifying parameter to size nodes by when size_parameters = "Node".

node_shape_by, a string specifying parameter to shape nodes by when shape_parameters = "Node".

order_tree, a logical indicating if tree is ordered by alphabetic order of taxonomic levels.

In addition, this class inherits all slots from its parent Panel class.

Author

Giulio Benedetti

Examples

# Import TreeSE
library(mia)
data("Tengeler2020", package = "mia")
tse <- Tengeler2020

# Launch iSEE
if (interactive()) {
  iSEE(tse)
}